Refactor, improvements, tweaks
