+++ app/javascript/components/modals/realtime-feedback-modal/CheckingForAutomatedFeedback.tsx
@@ -1,5 +1,7 @@
 import React, { memo } from 'react'
 import { GraphicalIcon } from '@/components/common'
+import { LoadingBar } from '@/components/common/LoadingBar'
+import { FooterButtonContainer } from './components'
 
 function CheckingForFeedback({
   onClick,
@@ -19,15 +21,11 @@ function CheckingForFeedback({
             Our systems are inspecting your code to find both automated feedback
             and feedback given by mentors on similar solutions.
           </p>
-          <p className="text-16 leading-150  mb-8">
+          <p className="text-16 leading-150  mb-12">
             This process normally takes ~10 seconds.
           </p>
-          <GraphicalIcon
-            icon="spinner"
-            className="animate-spin filter-textColor1"
-            height={24}
-            width={24}
-          />
+
+          <LoadingBar animationDuration={10} />
         </div>
         <GraphicalIcon
           height={160}
@@ -38,16 +36,18 @@ function CheckingForFeedback({
         />
       </div>
       {showTakingTooLong && <TakingTooLong />}
-      <button onClick={onClick} className="btn-secondary btn-s mr-auto">
-        Continue without waiting
-      </button>
+      <FooterButtonContainer>
+        <button onClick={onClick} className="btn-secondary btn-s mr-auto">
+          Continue without waiting
+        </button>
+      </FooterButtonContainer>
     </>
   )
 }
 
 function TakingTooLong(): JSX.Element {
   return (
-    <div class="c-textblock-caution">
+    <div className="c-textblock-caution mt-12 mb-20">
       <div className="c-textblock-content text-p-base leading-150">
         Sorry, this is taking a little longer than expected. You may wish to
         continue without waiting. You can view any feedback on the iterations
