+++ app/javascript/components/modals/realtime-feedback-modal/components/index.ts
@@ -1,0 +1,2 @@
+export { ContinueButton, GoBackToExercise } from './FeedbackContentButtons'
+export { FooterButtonContainer } from './FooterButtonContainer'
