+++ app/javascript/components/modals/realtime-feedback-modal/feedback-content/index.ts
@@ -1,3 +1,2 @@
-export { ContinueButton, GoBackToExercise } from './FeedbackContentButtons'
 export { NoAutomatedFeedback } from './no-automated-feedback'
 export { FoundAutomatedFeedback } from './found-automated-feedback'
