+++ app/javascript/components/modals/realtime-feedback-modal/hooks/useGetLatestIteration.ts
@@ -3,7 +3,7 @@ import { useQueryCache } from 'react-query'
 import { usePaginatedRequestQuery } from '@/hooks'
 import { SolutionWithLatestIterationChannel } from '@/channels/solutionWithLatestIterationChannel'
 import { IterationStatus } from '@/components/types'
-import type { RealtimeFeedbackModalProps, ResolvedIteration } from '.'
+import type { RealtimeFeedbackModalProps, ResolvedIteration } from '..'
 
 const REFETCH_INTERVAL = 2000
 const PENDING_STATUS = [
