+++ app/javascript/components/modals/realtime-feedback-modal/index.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react'
 import { redirectTo } from '@/utils'
 import { Modal } from '../Modal'
 import { FeedbackContent } from './FeedbackContent'
-import { useGetLatestIteration } from './useGetLatestIteration'
+import { useGetLatestIteration } from './hooks/useGetLatestIteration'
 import type { Props } from '@/components/editor/Props'
 import type { IterationsListRequest } from '@/components/student/IterationsList'
 import type { Submission } from '@/components/editor/types'
